requirements.txt
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pkg-resources==0.0.0
Werkzeug==1.0.0
